Added the class for compressed bitmap. Changed the cmake build to be 64bit. Moved to 64bit boost. Removed warnings for size_t and int mismatch. Updated the README.md file with new instructions.

+++ include/internal/document_collection.h
@@ -16,7 +16,7 @@ class DocumentCollection {
  public:
   ~DocumentCollection();
   static Status Construct(const char* databaseMetadataFilePath, const char* name, int schemaType, const char* schema,
-                          const IndexInfo indexes[], int indexesLength, DocumentCollection*& documentCollection);
+                          const IndexInfo indexes[], size_t indexesLength, DocumentCollection*& documentCollection);
   Status Insert(const Buffer& documentData);
 
  private:
