+++ src/jonoondb_api/index_manager.cc
@@ -10,7 +10,7 @@ using namespace jonoondb_api;
 IndexManager::IndexManager(const IndexInfo indexes[], int indexesLength) {
 }
 
-Status IndexManager::Construct(const IndexInfo indexes[], int indexesLength, IndexManager*& indexManager) {
+Status IndexManager::Construct(const IndexInfo indexes[], size_t indexesLength, IndexManager*& indexManager) {
   indexManager = new IndexManager(indexes, indexesLength);
   return Status();
 }
